[Refactor] Opt RuntimeFilter Phase2
