+++ be/src/exprs/runtime_filter_bank.h
@@ -39,12 +39,7 @@
 #include "runtime/runtime_state.h"
 #include "types/logical_type.h"
 
-namespace starrocks::pipeline {
-struct RuntimeMembershipFilterBuildParam;
-}
-
 namespace starrocks {
-struct SkewBroadcastRfMaterial;
 class RowDescriptor;
 class MemTracker;
 class ExecEnv;
@@ -54,38 +49,6 @@ class HashJoinNode;
 class RuntimeFilterProbeCollector;
 class RuntimeFilterHelper {
 public:
-    // ==================================
-    // serialization and deserialization.
-    static size_t max_runtime_filter_serialized_size(RuntimeState* state, const RuntimeFilter* rf);
-    static size_t max_runtime_filter_serialized_size(int rf_version, const RuntimeFilter* rf);
-    static size_t max_runtime_filter_serialized_size_for_skew_broadcast_join(const ColumnPtr& column);
-    static size_t serialize_runtime_filter(RuntimeState* state, const RuntimeFilter* rf, uint8_t* data);
-    static size_t serialize_runtime_filter(int rf_version, const RuntimeFilter* rf, uint8_t* data);
-    static StatusOr<size_t> serialize_runtime_filter_for_skew_broadcast_join(const ColumnPtr& column, bool eq_null,
-                                                                             uint8_t* data);
-    static int deserialize_runtime_filter(ObjectPool* pool, RuntimeFilter** rf, const uint8_t* data, size_t size);
-    static StatusOr<int> deserialize_runtime_filter_for_skew_broadcast_join(ObjectPool* pool,
-                                                                            SkewBroadcastRfMaterial** material,
-                                                                            const uint8_t* data, size_t size,
-                                                                            const PTypeDesc& ptype);
-
-    static RuntimeFilter* create_runtime_empty_filter(ObjectPool* pool, LogicalType type, int8_t join_mode);
-    static RuntimeFilter* create_runtime_bloom_filter(ObjectPool* pool, LogicalType type, int8_t join_mode);
-    static RuntimeFilter* create_runtime_bitset_filter(ObjectPool* pool, LogicalType type, int8_t join_mode);
-    static RuntimeFilter* create_agg_runtime_in_filter(ObjectPool* pool, LogicalType type, int8_t join_mode);
-    static RuntimeFilter* transmit_to_runtime_empty_filter(ObjectPool* pool, RuntimeFilter* rf);
-    static RuntimeFilter* create_runtime_filter(ObjectPool* pool, RuntimeFilterSerializeType rf_type, LogicalType ltype,
-                                                int8_t join_mode);
-    static RuntimeFilter* create_join_runtime_filter(ObjectPool* pool, LogicalType type, int8_t join_mode,
-                                                     const pipeline::RuntimeMembershipFilterBuildParam& param,
-                                                     size_t column_offset, size_t row_count);
-    // ====================================
-    static Status fill_runtime_filter(const ColumnPtr& column, LogicalType type, RuntimeFilter* filter,
-                                      size_t column_offset, bool eq_null, bool is_skew_join = false);
-    static Status fill_runtime_filter(const Columns& column, LogicalType type, RuntimeFilter* filter,
-                                      size_t column_offset, bool eq_null);
-    static Status fill_runtime_filter(const pipeline::RuntimeMembershipFilterBuildParam& param, LogicalType type,
-                                      RuntimeFilter* filter, size_t column_offset);
     static StatusOr<ExprContext*> rewrite_runtime_filter_in_cross_join_node(ObjectPool* pool, ExprContext* conjunct,
                                                                             Chunk* chunk);
