+++ be/src/runtime/runtime_filter_builder.h
@@ -1,0 +1,32 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "column/vectorized_fwd.h"
+#include "common/status.h"
+#include "runtime/runtime_filter.h"
+#include "types/logical_type.h"
+
+namespace starrocks {
+
+class RuntimeFilterBuilder {
+public:
+    static Status fill(RuntimeFilter* filter, LogicalType type, const ColumnPtr& column, size_t column_offset,
+                       bool eq_null, bool is_skew_join = false);
+    static Status fill(RuntimeFilter* filter, LogicalType type, const Columns& columns, size_t column_offset,
+                       bool eq_null, bool is_skew_join = false);
+};
+
+} // namespace starrocks
