+++ be/src/runtime/runtime_filter_worker.h
@@ -32,6 +32,7 @@
 #include "gen_cpp/InternalService_types.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
+#include "runtime/runtime_filter_serde.h"
 namespace starrocks {
 struct TypeDescriptor;
 
@@ -73,11 +74,7 @@ private:
     RuntimeState* _state;
 };
 
-struct SkewBroadcastRfMaterial {
-    LogicalType build_type;
-    bool eq_null;
-    ColumnPtr key_column;
-};
+using SkewBroadcastRfMaterial = RuntimeFilterSkewMaterial;
 
 class RuntimeFilterMergerStatus {
 public:
